Rename finder to find
