+++ internal/find/find.go
@@ -1,4 +1,4 @@
-package finder
+package find
 
 import (
 	"fmt"
